requirements.txt
flet
flet-charts
mysql-connector-python
Pillow
python-dotenv